The spatial locations of the Beijing land price data

The spatial locations of the Beijing land price data

Usage

land

Format

An object of class sf (inherits from data.frame) with 1117 rows and 3 columns.