Leased residential land parcels, from 2003 to 2009 in Beijing, China

Leased residential land parcels, from 2003 to 2009 in Beijing, China

Usage

landprice

Format

An object of class data.frame with 1117 rows and 11 columns.

Details

A data.frame with 1117 observations on the following 11 variables.

obs

An unique identifier for each land parcel.

lnprice

The log of the leasing price per square metre of each residential land parcel (unit: RMB, Chinese yuan)

dsubway

The log of the distance of each land parcel to the nearest railway station (unit:meters)

dele

The log of the distance of each land parcel to the nearest elementary school (unit:meters)

dpark

The log of the distance of each land parcel to the nearest green park (unit:meters)

lnarea

The log of the size of each land parcel (unit: square meters).

lndcbd

The log of the distance of each land parcel to the CBD (centre business district) in Beijing (unit:meters)

year

The year when each land parcel was leased with values of 0,1,2,3,4,5,6 representing year 2003,2004,2005,2006,2007,2008,2009

popden

The population density of each district (unit: 1000 persons per square kilometers)

crimerate

The number of reported serious crimes committed in each district per 1000 persons.

district.id

The identifier of the district where each land parcel is located.